+++ CHANGES.rst
@@ -28,6 +28,8 @@ Changelog
   [chris-adam, sgeulette]
 - Fixed AttributeError (get_person) raised in is_hp_used_in_signer_rules (PARAF-477).
   [chris-adam]
+- Allow removal of userid on a person if he's not in any group.
+  [chris-adam]
 
 3.1.4 (2026-05-29)
 ------------------